Data Polling
This creates timers which pull the database to keep the views fresh.
This presents a unique challenge as the timers must be properly
initiated and cleared when the nav state changes or the app state
changes. This code is horrible and brittle but I think it can be
refactored into some kind of mixin like module which would get rid of
all this code noise as well as simplify the state. There will almost
certainly be race conditions we need to look out for. Also, I had to
include some experimental code for react-navigation in order to use the
navigation state changes listeners. Why that is not a thing yet, i have
no idea.